requirements.txt
# Web Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text Processing
tiktoken>=0.5.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings and Vector Store
openai>=1.12.0
chromadb>=0.4.22

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
validators>=0.22.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
